[CHORE] 더미 데이터 파일 추가
